+++ game_sessions/usecase.go
@@ -5,5 +5,5 @@ import (
 )
 
 type UseCase interface {
-	NewSession(ctx context.Context, playerId uint32) error
+	NewSession(ctx context.Context, playerId uint64) error
 }
